feat: fsh-api-cli tool
