+++ FSH.WebApi.nuspec
@@ -11,6 +11,7 @@
     <language>en-US</language>
     <license type="expression">MIT</license>
     <copyright>2023</copyright>
+    <readme>README.md</readme>
     <projectUrl>https://fullstackhero.net/dotnet-webapi-boilerplate/general/getting-started/</projectUrl>
     <repository type="git" url="https://github.com/fullstackhero/dotnet-webapi-boilerplate.git"/>
     <packageTypes>
@@ -21,5 +22,6 @@
   </metadata>
   <files>
     <file src=".\**" target="content" exclude="**\bin\**;**\obj\**;**\.git\**;**\.github\**;**\*.user;**\.vs\**;**\Images\**;**\Logs\**" />
+    <file src="README.md" target="\" />
   </files>
 </package>
