+++ fsh-api-cli/FSH.API.CLI.csproj
@@ -1,5 +1,4 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net7.0</TargetFramework>
@@ -7,13 +6,12 @@
     <ToolCommandName>fsh-api</ToolCommandName>
     <PackageOutputPath>./nupkg</PackageOutputPath>
     <PackageVersion>1.0.0</PackageVersion>
-    <PackageId>FSH.API.Console</PackageId>
-    <Title>FSH API CLI</Title>
+    <PackageId>FSH.API.CLI</PackageId>
+    <Title>FullStackHero .NET WebAPI CLI Tool</Title>
     <Authors>Mukesh Murugan</Authors>
     <Description>FSH CLI tool for creating and using FSH .NET WebAPI Boilerplate Projects.</Description>
     <PackageTags>dotnet-new;templates</PackageTags>
     <RepositoryUrl>https://github.com/fullstackhero/dotnet-webapi-boilerplate</RepositoryUrl>
     <PackageProjectUrl>https://fullstackhero.net/</PackageProjectUrl>
   </PropertyGroup>
-
 </Project>
