+++ fsh-api-cli/README.md
@@ -1,0 +1,3 @@
+dotnet pack
+dotnet tool install --global --add-source ./nupkg FSH.API.CLI --version 1.0.0
+dotnet tool uninstall --global FSH.API.CLI
