--- fsh-api-console/README.md
@@ -1,3 +1,0 @@
-dotnet pack
-dotnet tool install --global --add-source ./nupkg FSH.API.Console --version 1.0.0
-dotnet tool uninstall --global FSH.API.Console
